requirements.txt
solders==0.18.1
base58==2.1.1
requests==2.31.0
colorama==0.4.6
python-dotenv==1.0.0
pynacl==1.5.0 